Refactored IPv8 class to an importable location
